resources/requirements.txt
ansible==2.8.5
colorama~=0.4.3
paramiko~=2.7.1
pytest~=6.0.1
pyvmomi==6.7.1.2018.12
pywinrm==0.4.1
tox==3.7.0
veryprettytable==0.8.1

requests~=2.22.0
elasticsearch~=7.14.0
elasticsearch_dsl~=7.4.0
python-dateutil~=2.7.3
setuptools~=45.2.0